+++ FXTracker/src/main/java/com/FXTracker/model/Portfolio.java
@@ -10,6 +10,10 @@ import org.springframework.data.mongodb.core.mapping.Field;
 
 import java.util.Map;
 
+/**
+ * Represents User Portfolio with details such as id, user id, stocks, balance, profit, and loss.
+ */
+
 @Data
 @Builder
 @NoArgsConstructor
@@ -24,9 +28,9 @@ public class Portfolio {
     @Field(name = "stocks")
     private Map<String, String> stocks;
     @Field(name = "balance")
-    private Float balance;
+    private Double balance;
     @Field(name = "profit")
-    private Float profit;
+    private Double profit;
     @Field(name = "loss")
-    private Float loss;
+    private Double loss;
 }
